requirements.txt
transformers==4.6.1
redis==3.5.3
psycopg2-binary==2.8.6
peewee==3.14.4
scipy==1.6.3